Clean up python code
